added this for generation of different name for the file
Generating a UUID (aka GUID) in Java.

What are the practical use cases for UUID?

As a row identity in a database table. i.e. as a primary key. The
benefit is that it will be unique across every table, every database,
and every server. You can generate IDs anywhere. The disadvantage is
that it is 4 times larger than the traditional 4-byte index value. UUIDs
may not index well and could degrade database performance.
UUID is a natural fit for many development scenarios where you need to
generate primary keys outside the database.
Many data replication scenarios require UUID columns. Another option is
to let the database generate a sequence key for internal relationships,
and use a UUID as a unique secondary key.
It can be used as an audit tracking id in web service calls.
It can also be used as a message correlation id.
Temporary file names.
Transaction IDs.
Unique identifiers for website users/visitors.