ref: http://www.thoughts-on-java.org/how-to-activate-hibernate-statistics-to-analyze-performance-issues/
If you need to analyze performance issues, Hibernate can collect and log
useful statistics about each database query and the whole session. This
needs to be activated by setting the system property
hibernate.generate_statistics to true and the log level for
org.hibernate.stat to DEBUG.

You can then decide, if the performance issue is caused by the database
communication or by some other part of the system. If it is caused by
the database communication, you can see if you need to fix it in your
Java code or if you need to contact your database administrator to get
more detailed information about your slow queries.